refactor: done part2
