+++ lib/features/settings/categories/app.dart
