+++ lib/features/settings/categories/components/body_listcate.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:testverygood/components/Ex_In_btn_Satis.dart';
-import 'package:testverygood/features/settings/add_categories/components/content.dart';
-import 'package:testverygood/features/settings/add_categories/view/addcate_main.dart';
+import 'package:testverygood/features/settings/categories/components/content.dart';
+import 'package:testverygood/features/settings/categories/view/addcate_main.dart';
 
 class BodyMain extends StatefulWidget {
   const BodyMain({super.key});
