+++ lib/features/settings/categories/view/addcate_main.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:testverygood/components/HeaderA.dart';
-import 'package:testverygood/features/settings/add_friends/components/body_groups.dart';
+import 'package:testverygood/features/settings/categories/components/body_addcate.dart';
 
-class AddGroupPage extends StatelessWidget {
-  const AddGroupPage({super.key});
+class AddCateMain extends StatelessWidget {
+  const AddCateMain({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      backgroundColor: Colors.white,
-      appBar: HeaderA(title: 'Add group'),
+      appBar: HeaderA(title: 'Categories'),
       body: Padding(
-        padding: EdgeInsets.zero,
+        padding: EdgeInsets.all(20),
         child: Expanded(
           child: BodyMain(),
         ),
