+++ lib/features/settings/groupfriends/view/friendlist_main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:testverygood/components/HeaderA.dart';
-import 'package:testverygood/features/settings/add_friends/components/body_friends.dart';
+import 'package:testverygood/features/settings/groupfriends/components/body_friends.dart';
 
 class FriendListPage extends StatelessWidget {
   const FriendListPage({super.key});
