+++ lib/features/transaction/add_trans/components/categories.dart
