+++ lib/features/transaction/add_trans/view/transaction_main.dart
@@ -5,9 +5,9 @@ import 'package:flutter_svg/svg.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:testverygood/components/HeaderA.dart';
 import 'package:testverygood/components/input.dart';
-import 'package:testverygood/features/transactions/transactrion/components/Ex_In_btn.dart';
-import 'package:testverygood/features/transactions/transactrion/components/calendar.dart';
-import 'package:testverygood/features/transactions/transactrion/components/categories.dart';
+import 'package:testverygood/features/transaction/add_trans/components/Ex_In_btn.dart';
+import 'package:testverygood/features/transaction/add_trans/components/calendar.dart';
+import 'package:testverygood/features/transaction/add_trans/components/categories.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:testverygood/features/main_navbar.dart';
 import 'package:testverygood/data/data_api/add_trans.dart';
